backend/requirements.txt
# FastAPI Backend Requirements for PLN Ticket System

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# CORS
fastapi-cors==0.0.6

# Logging & Monitoring
python-json-logger==2.0.7

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0